Rename component in date to relative-date
